Fixes (Song resync, setting rounding errors + more) (#44)
* fadeout adjustments

* scale song pitch when desynced from map

* reallow window resize

* safer pitch scaling conditions

* min pitch scale > 0

* decouple settings menu toggling from settingsmanager (+fix double esc to open menu in gameplay after opening settings)

* oopsie!!

* fix: progress bar incremental delay

* placeholder versioning

* fix: negative note alpha in replays

* fix: clicking outside update popup softlocks

* fix: displayed selected profile not accurate

* fix: max cursor opacity disables alpha support

* fix: ranged settings rounding error & imprecision

* space scripts cleanup

* space redundancies

+++ scripts/KeybindsManager.cs
@@ -47,7 +47,7 @@ public partial class KeybindsManager : Node
             {
                 if (volumeBeforeMute < 0)
                 {
-                    volumeBeforeMute = settings.VolumeMaster.Value;
+                    volumeBeforeMute = (float)settings.VolumeMaster.Value;
                     settings.VolumeMaster.Value = 0;
                 }
                 else
